traz resumo de anotações para página principal
