+++ frontend/src/router/tasks.routes.js
@@ -1,7 +1,6 @@
 // Composables
 import DefaultLayout from "@/layouts/default/DefaultLayout.vue"
 import TaskListView from "@/views/tasks/TaskListView.vue"
-import TaskNovaView from "@/views/tasks/TaskNovaView.vue"
 
 export default [
   {
@@ -12,11 +11,6 @@ export default [
         path: "list",
         name: "tasks-list",
         component: TaskListView,
-      },
-      {
-        path: "novanotacao",
-        name: "tasks-nova",
-        component: TaskNovaView,
       },
     ],
   },
